// src/gen/types/GetApiClientMeRoutineWorkouts.ts
/**
* Generated by Kubb (https://kubb.dev/).
* Do not edit manually.
*/


/**
 * @description Response for status 200
*/
export type GetApiClientMeRoutineWorkouts200 = {
    /**
     * @type array
    */
    workouts: {
        /**
         * @type string
        */
        id: string;
        /**
         * @type string
        */
        name: string;
        /**
         * @type string | undefined
        */
        description?: string;
        ownerId?: (string | null);
        /**
         * @type string
        */
        routineId: string;
        /**
         * @minLength -9007199254740991
         * @maxLength 9007199254740991
         * @type integer
        */
        order: number;
        /**
         * @type array | undefined
        */
        exercises?: {
            /**
             * @type string
            */
            id: string;
            /**
             * @minLength -9007199254740991
             * @maxLength 9007199254740991
             * @type integer
            */
            order: number;
            /**
             * @type array
            */
            sets: {
                /**
                 * @type string | undefined
                */
                type?: string;
                /**
                 * @maxLength 9007199254740991
                 * @type integer | undefined
                */
                reps?: number;
                /**
                 * @maxLength 9007199254740991
                 * @type integer | undefined
                */
                weight?: number;
                /**
                 * @maxLength 9007199254740991
                 * @type integer | undefined
                */
                rest?: number;
            }[];
            /**
             * @type object
            */
            exerciseData: {
                /**
                 * @type string
                */
                id: string;
                /**
                 * @type string
                */
                exerciseId: string;
                /**
                 * @type string
                */
                name: string;
                /**
                 * @type string
                */
                category: string;
                imgSrc?: (string | null);
                /**
                 * @type string
                */
                equipment: string;
                /**
                 * @type string
                */
                primaryMuscle: string;
                /**
                 * @type string
                */
                secondaryMuscle: string;
                howTo?: (string | null);
                videoUrl?: (string | null);
                ownerId?: (string | null);
            };
        }[];
        /**
         * @type string
        */
        createdAt: string;
        /**
         * @type string
        */
        updatedAt: string;
    }[];
};

/**
 * @description Response for status 403
*/
export type GetApiClientMeRoutineWorkouts403 = {
    /**
     * @type string
    */
    message: string;
};

/**
 * @description Response for status 404
*/
export type GetApiClientMeRoutineWorkouts404 = {
    /**
     * @type string
    */
    message: string;
};

export type GetApiClientMeRoutineWorkoutsQueryResponse = GetApiClientMeRoutineWorkouts200;

export type GetApiClientMeRoutineWorkoutsQuery = {
    Response: GetApiClientMeRoutineWorkouts200;
    Errors: GetApiClientMeRoutineWorkouts403 | GetApiClientMeRoutineWorkouts404;
};